Fasttrack: extract visuals to separate file
